+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * `Icon` factories now support `intent`.
 * `TreeMapModel` and `SplitTreeMapModel` now supports a `theme` config, accepting the strings
   'light' or 'dark'. Leave it undefined to use the global theme.
+### 🐞 Bug Fixes
+* Fixed an issue preventing `FormField` labels from rendering if `fieldDefaults` was undefined.
 
 ### ✨ Style
 * New `compact` prop on `Badge` component sets size to half that of parent element when true,